refactor: API, HTML 템플릿 파일 분리
