+++ src/js/components/Map.js
@@ -1,42 +1,10 @@
 import Component from "./common/Component";
 import { getSessionStorageItem } from "../utils/sessionStorage.js";
-import { getLinesAPI } from "../APIs/subwayAPI.js";
+import { getLinesAPI } from "../APIs/index.js";
 import { TOKEN_STORAGE_KEY } from "../constants/general.js";
 import { PAGE_KEYS, PAGE_URLS } from "../constants/pages.js";
+import { createLineTemplate } from "../constants/template.js";
 import { $, removeAllChildren } from "../utils/DOM.js";
-
-const createSectionTemplate = (station, section = {}, color) => {
-  return `
-    <li class="map-station"><span class="${color}"></span>${station.name}</li>
-    ${
-      section.distance
-        ? `<li class="map-section-info">
-      시간 : ${section.duration} / 거리 : ${section.distance}
-    </li>`
-        : ""
-    }
-  `;
-};
-
-const createLineTemplate = (line) => {
-  const sections = {};
-  line.sections.forEach((section) => {
-    sections[section.upStation.name] = section;
-  });
-
-  return `
-    <section class="js-map-line-wrapper">
-      <h3 class="map-line-name mt-3">
-        ${line.name}
-      </h3>
-      <ol class="map-line-list">
-        ${line.stations
-          .map((v) => createSectionTemplate(v, sections[v.name], line.color))
-          .join("")}
-      </ol>
-    </section>
-  `;
-};
 
 export default class Map extends Component {
   constructor({ $parent, setPageState }) {
